fix: catch-all paths when directly accessing project
Directly accessing a project into a fullscreen modal doesn't work atm.
Modal works and opens a routed path because it uses location history. If
trying to open a modal directly, no history for it to know what it's
previous page/action was - aka parent project and clicking to open the
modal. If I can find a way to circumvent this, then I can remove these
catch-all routes.